Split mapblock mesh into several files
