+++ src/wieldmesh.cpp
@@ -25,7 +25,8 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "itemdef.h"
 #include "nodedef.h"
 #include "mesh.h"
-#include "mapblock_mesh.h"
+#include "meshgen/params.h"
+#include "meshgen/mapblock_mesh.h"
 #include "client/tile.h"
 #include "log.h"
 #include "util/numeric.h"
